+++ packages/cli/src/utils/contracts.js
@@ -1,4 +1,5 @@
 import crypto from 'crypto'
+import flatten from 'truffle-flattener'
 
 export function bodyCode(instance) {
   return splitCode(instance).body
@@ -20,4 +21,8 @@ export function bytecodeDigest(rawBytecode) {
   const buffer = Buffer.from(bytecode, 'hex')
   const hash = crypto.createHash('sha256')
   return hash.update(buffer).digest('hex')
+}
+
+export function flattenSourceCode(contract) {
+  return flatten(contract)
 }